feat: Add JsonPropertyOrder annotation to all classes
Stage 1 of Issue 2779
Validate @JsonPropertyOrder present for all CRD classes
and subclasses with exception of Java primitive `Object`

Added test class to validate this new functionality and
added Exception: InvalidCrdException

Followup PRs will see the check that JsonPropertyOrder
contains exactly all of the classes fields and these fields
will be ordered for each individual class

Contributes to: #2779
