+++ api/src/main/java/io/strimzi/api/kafka/model/authentication/KafkaClientAuthentication.java
@@ -5,6 +5,7 @@
 package io.strimzi.api.kafka.model.authentication;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.strimzi.api.kafka.model.UnknownPropertyPreserving;
@@ -28,6 +29,7 @@ import java.util.Map;
         @JsonSubTypes.Type(name = KafkaClientAuthenticationOAuth.TYPE_OAUTH, value = KafkaClientAuthenticationOAuth.class),
 })
 @JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({})
 @EqualsAndHashCode
 public abstract class KafkaClientAuthentication implements UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
