+++ api/src/main/java/io/strimzi/api/kafka/model/listener/KafkaListenerAuthentication.java
@@ -5,6 +5,7 @@
 package io.strimzi.api.kafka.model.listener;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.strimzi.api.kafka.model.Constants;
@@ -33,6 +34,7 @@ import java.util.Map;
         editableEnabled = false,
         builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
+@JsonPropertyOrder({})
 @EqualsAndHashCode
 public abstract class KafkaListenerAuthentication implements UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
