+++ api/src/main/java/io/strimzi/api/kafka/model/listener/NodePortListenerBootstrapOverride.java
@@ -5,6 +5,7 @@
 package io.strimzi.api.kafka.model.listener;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import io.strimzi.api.kafka.model.Constants;
 import io.strimzi.crdgenerator.annotations.Description;
 import io.sundr.builder.annotations.Buildable;
@@ -24,6 +25,7 @@ import static java.util.Collections.emptyMap;
     editableEnabled = false,
     builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
+@JsonPropertyOrder({})
 @EqualsAndHashCode(callSuper = true)
 public class NodePortListenerBootstrapOverride extends ExternalListenerBootstrapOverride {
     private static final long serialVersionUID = 1L;
