+++ crd-generator/src/main/java/io/strimzi/crdgenerator/InvalidCrdException.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.crdgenerator;
+
+public class InvalidCrdException extends RuntimeException {
+
+    public InvalidCrdException(String message) {
+        super(message);
+    }
+}
